+++ src/BigNumber.php
@@ -360,6 +360,39 @@ abstract readonly class BigNumber implements JsonSerializable, Stringable
     abstract public function compareTo(BigNumber|int|float|string $that): int;
 
     /**
+     * Limits (clamps) this number between the given minimum and maximum values.
+     *
+     * If the number is lower than $min, returns a copy of $min.
+     * If the number is greater than $max, returns a copy of $max.
+     * Otherwise, returns this number unchanged.
+     *
+     * @param BigNumber|int|float|string $min The minimum. Must be convertible to an instance of the class this method is called on.
+     * @param BigNumber|int|float|string $max The maximum. Must be convertible to an instance of the class this method is called on.
+     *
+     * @throws MathException            If min/max are not convertible to an instance of the class this method is called on.
+     * @throws InvalidArgumentException If min is greater than max.
+     *
+     * @pure
+     */
+    public function clamp(BigNumber|int|float|string $min, BigNumber|int|float|string $max): static
+    {
+        $min = static::of($min);
+        $max = static::of($max);
+
+        if ($min->isGreaterThan($max)) {
+            throw new InvalidArgumentException('Minimum value must be less than or equal to maximum value.');
+        }
+
+        if ($this->isLessThan($min)) {
+            return $min;
+        } elseif ($this->isGreaterThan($max)) {
+            return $max;
+        }
+
+        return $this;
+    }
+
+    /**
      * Converts this number to a BigInteger.
      *
      * @throws RoundingNecessaryException If this number cannot be converted to a BigInteger without rounding.