+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED (0.14.7)
+
+✨ **New features**
+
+- `clamp()` is now available on the base `BigNumber` class
+
 ## [0.14.6](https://github.com/brick/math/releases/tag/0.14.6) - 2026-02-05
 
 🗑️ **Deprecations**